+++ examples/grpc.core/Examples.GrpcCore.AspNetCore/Controllers/WeatherForecastController.cs
@@ -7,12 +7,14 @@ namespace Examples.GrpcCore.AspNetCore.Controllers;
 
 [ApiController]
 [Route("[controller]")]
+#pragma warning disable CA1515
 public class WeatherForecastController : ControllerBase
+#pragma warning restore CA1515
 {
-    private static readonly string[] Summaries = new[]
-    {
-        "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching",
-    };
+    private static readonly string[] Summaries =
+    [
+        "Freezing", "Bracing", "Chilly", "Cool", "Mild", "Warm", "Balmy", "Hot", "Sweltering", "Scorching"
+    ];
 
     private readonly Echo.EchoClient echoClient;