+++ src/OpenTelemetry.Exporter.Geneva/Internal/Transports/EtwDataTransport.cs
@@ -67,11 +67,11 @@ internal sealed class EtwDataTransport : IDataTransport, IDisposable
 #endif
         public unsafe void SendEvent(int eventId, byte[] data, int size)
         {
-            EventData* dataDesc = stackalloc EventData[1];
+            var dataDesc = stackalloc EventData[1];
             fixed (byte* ptr = data)
             {
                 dataDesc[0].DataPointer = (IntPtr)ptr;
-                dataDesc[0].Size = (int)size;
+                dataDesc[0].Size = size;
                 this.WriteEventCore(eventId, 1, dataDesc);
             }
         }