+++ test/OpenTelemetry.Exporter.Geneva.Benchmarks/Exporter/SerializationBenchmarks.cs
@@ -32,7 +32,9 @@ Intel Core i7-9700 CPU 3.00GHz, 1 CPU, 8 logical and 8 physical cores
 namespace OpenTelemetry.Exporter.Geneva.Benchmarks;
 
 [MemoryDiagnoser]
+#pragma warning disable CA1515
 public class SerializationBenchmarks
+#pragma warning restore CA1515
 {
     private const int StringLengthLimit = (1 << 14) - 1;
     private const string Key = "ext_dt_traceId";