+++ test/OpenTelemetry.Exporter.Geneva.Benchmarks/Exporter/TLDTraceExporterBenchmarks.cs
@@ -26,13 +26,15 @@ Intel Core i7-9700 CPU 3.00GHz, 1 CPU, 8 logical and 8 physical cores
 namespace OpenTelemetry.Exporter.Geneva.Benchmarks;
 
 [MemoryDiagnoser]
+#pragma warning disable CA1515
 public class TLDTraceExporterBenchmarks
+#pragma warning restore CA1515
 {
     private readonly Activity? activity;
     private readonly Batch<Activity> batch;
     private readonly MsgPackTraceExporter msgPackExporter;
     private readonly TldTraceExporter tldExporter;
-    private readonly ActivitySource activitySource = new ActivitySource("OpenTelemetry.Exporter.Geneva.Benchmark");
+    private readonly ActivitySource activitySource = new("OpenTelemetry.Exporter.Geneva.Benchmark");
 
     public TLDTraceExporterBenchmarks()
     {