+++ src/OpenTelemetry.Instrumentation.AspNet/CHANGELOG.md
@@ -9,6 +9,10 @@
 * Updated OpenTelemetry core component version(s) to `1.10.0`.
   ([#2317](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2317))
 
+* Fixed an issue in ASP.NET instrumentation where route extraction failed for
+  attribute-based routing with multiple HTTP methods sharing the same route template.
+  ([#2250](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2250))
+
 ## 1.9.0-beta.1
 
 Released 2024-Jun-18