+++ src/OpenTelemetry.Instrumentation.Http/CHANGELOG.md
@@ -2,11 +2,20 @@
 
 ## Unreleased
 
+## 1.10.0
+
+Released 2024-Nov-27
+
 * Drop support for .NET 6 as this target is no longer supported.
   ([#2152](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2152))
 
 * Updated OpenTelemetry core component version(s) to `1.10.0`.
   ([#2317](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2317))
+
+* Trace instrumentation no longer sets attributes when running on .NET 9 and
+  greater because `HttpClient` now includes native instrumentation which adds
+  attributes directly.
+  ([#2314](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/2314))
 
 ## 1.9.0